+++ gol-engine/src/universe/sparse_grid/mod.rs
@@ -4,7 +4,7 @@ use std::collections::{HashMap, HashSet};
 use std::iter::Iterator;
 
 pub struct SparseGrid {
-    grid: HashMap<i8, HashSet<i8>>,
+    grid: HashMap<isize, HashSet<isize>>,
 }
 
 impl SparseGrid {
@@ -20,14 +20,14 @@ impl SparseGrid {
         }
     }
 
-    pub fn has_value(&self, row_index: i8, column_index: i8) -> bool {
+    pub fn has_value(&self, row_index: isize, column_index: isize) -> bool {
         self.grid
             .get(&row_index)
             .map(|row| row.contains(&column_index))
             .unwrap_or_default()
     }
 
-    pub fn append_value(&mut self, row_index: i8, column_index: i8) -> () {
+    pub fn append_value(&mut self, row_index: isize, column_index: isize) -> () {
         let row = self.grid.get_mut(&row_index);
         match row {
             Some(row) => {
@@ -53,11 +53,11 @@ impl SparseGrid {
 }
 
 pub struct SparseGridIter<'a> {
-    inner_iter: Box<dyn Iterator<Item = (i8, i8)> + 'a>,
+    inner_iter: Box<dyn Iterator<Item = (isize, isize)> + 'a>,
 }
 
 impl<'a> Iterator for SparseGridIter<'a> {
-    type Item = (i8, i8);
+    type Item = (isize, isize);
 
     fn next(&mut self) -> Option<Self::Item> {
         self.inner_iter.next()
